refactor(metadata): migrate inventoryhost component to v2
- component.go -> def/component.go
- inventoryhostimpl/ -> impl/
- Mock moved to mock/mock.go
- fx.In/Out -> plain Requires/Provides structs
- Update all callers to use new def and fx packages
