+++ cmd/agent/subcommands/flare/command.go
@@ -60,7 +60,7 @@ import (
 	haagentmetadatafx "github.com/DataDog/datadog-agent/comp/metadata/haagent/fx"
 	hostfx "github.com/DataDog/datadog-agent/comp/metadata/host/fx"
 	inventoryagentfx "github.com/DataDog/datadog-agent/comp/metadata/inventoryagent/fx"
-	"github.com/DataDog/datadog-agent/comp/metadata/inventoryhost/inventoryhostimpl"
+	inventoryhostfx "github.com/DataDog/datadog-agent/comp/metadata/inventoryhost/fx"
 	resourcesfx "github.com/DataDog/datadog-agent/comp/metadata/resources/fx"
 	logscompressorfx "github.com/DataDog/datadog-agent/comp/serializer/logscompression/fx"
 	metricscompressorfx "github.com/DataDog/datadog-agent/comp/serializer/metricscompression/fx"
@@ -160,7 +160,7 @@ func Commands(globalParams *command.GlobalParams) []*cobra.Command {
 				// We need inventoryagent to fill the status page generated by the flare.
 				inventoryagentfx.Module(),
 				hostfx.Module(),
-				inventoryhostimpl.Module(),
+				inventoryhostfx.Module(),
 				haagentmetadatafx.Module(),
 				resourcesfx.Module(),
 				// inventoryagent require a serializer. Since we're not actually sending the payload to
