+++ comp/metadata/inventoryhost/def/component.go
