+++ comp/metadata/inventoryhost/fx/fx.go
@@ -1,0 +1,18 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+// Package fx provides the fx module for the inventoryhost metadata component
+package fx
+
+import (
+	inventoryhostimpl "github.com/DataDog/datadog-agent/comp/metadata/inventoryhost/impl"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+)
+
+// Module defines the fx options for this component.
+func Module() fxutil.Module {
+	return fxutil.Component(
+		fxutil.ProvideComponentConstructor(inventoryhostimpl.NewComponent))
+}
